Split dashboard Labs widget renderers
