+++ tests/test-correctness-phase2.js
@@ -124,6 +124,7 @@ const pwaAppShellAssets = [
   '/js/lens-page-shell.js',
   '/js/dashboard-widgets.js',
   '/js/dashboard-widget-controls.js',
+  '/js/dashboard-lab-widget-renderers.js',
   '/js/dashboard-widget-renderers.js',
   '/js/dashboard-recommendation-widget.js',
   '/js/recommendation-actions.js',
