+++ tsconfig.checkjs.json
@@ -87,6 +87,7 @@
     "js/context-cards.js",
     "js/crypto.js",
     "js/dashboard-page-view.js",
+    "js/dashboard-lab-widget-renderers.js",
     "js/dashboard-recommendation-widget.js",
     "js/dashboard-view-composition.js",
     "js/dashboard-widget-runtime.js",
